add get_specs.sh to regenerate Packages-puppy-noarch-official
The idea of this is for it to be sourced in woof-CE and the version
in woof deleted. That will save making extra commits to woof.
Just add/change/delete your packages(s) and run the script. Locale
is no issue as LC_COLLATE=C is set for sort. A new
Packages-puppy-noarch-official file is generated. The script is clickable!
I'll add something to the readme.md